requirements.txt
streamlit
python-docx
PyPDF2
langchain
langchain-google-genai
python-dotenv
pytesseract
pdf2image 
pillow